pyproject.toml
[tool.poetry]
name = "slack-twt"
version = "0.1.0"
description = "TWT Python Slack Bot Tutorial"
authors = ["rootshellz <rootshellz@rootshellz.com>"]

[tool.poetry.dependencies]
python = "^3.9"
slackclient = "^2.9.4"
python-dotenv = "^0.20.0"
Flask = "^2.1.3"
slackeventsapi = "^3.0.1"

[tool.poetry.dev-dependencies]
bpython = "^0.22.1"
black = "^22.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
